feat: Fetch and display stock data from 'https://dumbstockapi.com'
This commit adds a React component that fetches stock data from 'https://dumbstockapi.com/stock?exchanges=NYSE' using the 'axios' library. The fetched data is displayed in a list format on the page.

The component uses the 'useState' and 'useEffect' hooks to manage the data state and loading state. Initially, the 'loading' state is set to 'true' to display a loading message while the data is being fetched. Once the data is successfully fetched, the 'loading' state is set to 'false', and the retrieved stock information is displayed in an unordered list.

In case of an error during the data fetching process, an error message will be logged to the console.

The 'useEffect' hook ensures that the data is fetched when the component is mounted, and it also handles cleanup by returning a function that can be executed when the component is unmounted.

Note: The 'useEffect' hook has a commented-out part that shows where you can add a function to cancel the request in case the component is unmounted before the data is fetched. You can uncomment and implement that functionality if necessary.

To use this component, import it into your application and include it in the desired part of your UI.